7i.c: fix segment limits
Segment boundaries were incorrectly alligned at 1MB.  Only the text
and data segment base needs to be 1MB-alligned.

This causes a very hard to debug big when accessing data from what
should have been BSS.

The problem was that we got an emulated page-in in vaddr.  We read the
data page from the file (because it matched the data, not bss limits),
but it's at the end of the data segment, so code in vaddr will want to
zero out the trailing bytes.  This code assumes we didn't read much
past the data segment; at most up to the next page.

This assumption is generally correct, if we would have tried to read
further, we'd have faulted already.  In our case, however, we didn't
set the segment size correctly.  We alligned it at 1MB, but it should
be alligned at 4kB.  Only the segment *base* should be 1MB-alligned.

Because the segment end was much further than it should have been, and
the vaddr code didn't fault early, we ended up doing a memset that
overrided adjacent memory, in our case it zeroed out part of a text
page.

I was seeing this text corruption, so I tried to identify when it
happens by poking at the memory with getmem_w.  This caused the
page-in to happen much earlier in the program timeline, and while the
memset was still overwriting memory, it didn't corrupt the text pages
in the same way and we didn't see the problem.

After some time I started to poke at the memory directly, instead of
using getmem_w.  This preserved the allocation pattern, and preserved
the bug's behavior.  After that, I quickly identified where the
problem occurs and what causes it.